♻️ refactor: groupByDate utils에 분리
