+++ src/app/page.tsx
@@ -12,36 +12,10 @@ import {
 } from '@/components';
 import { DUMMY_IMAGE } from '@/constants';
 import { ThumbnailPhotoData } from '@/types';
-import dayjs from 'dayjs';
 import { toast } from 'react-toastify';
 import { useInfiniteScroll } from '@/hooks';
 import { RefObject, useEffect, useState } from 'react';
-
-interface DateGroup {
-  date: string;
-  photos: ThumbnailPhotoData[];
-}
-
-const groupByDate = (photos: ThumbnailPhotoData[]): DateGroup[] => {
-  const groupedMap: { [key: string]: ThumbnailPhotoData[] } = {};
-  const groupedArray: DateGroup[] = [];
-
-  photos.forEach(photo => {
-    const date = dayjs(photo.uploadDateTime).format('YYYY년 M월 D일');
-
-    if (!groupedMap[date]) {
-      groupedMap[date] = [];
-    }
-
-    groupedMap[date].push(photo);
-  });
-
-  for (const date in groupedMap) {
-    groupedArray.push({ date, photos: groupedMap[date] });
-  }
-
-  return groupedArray;
-};
+import { groupByDate, DateGroup } from '@/utils';
 
 export default function Home() {
   const postImageUpload = useUploadImage();
